[language] rename vm-runtime to libra-vm
The Move VM has been split out of this crate, so what remains is exactly the
Libra-specific bits of the VM.
