+++ Cargo.toml
@@ -54,6 +54,7 @@ members = [
     "language/tools/test-generation",
     "language/tools/vm-genesis",
     "language/compiler/bytecode-source-map",
+    "language/libra-vm",
     "language/move-ir/types",
     "language/move-prover/bytecode-to-boogie",
     "language/move-prover/spec-lang",
@@ -66,7 +67,6 @@ members = [
     "language/stdlib",
     "language/vm",
     "language/vm/serializer_tests",
-    "language/vm/vm-runtime",
     "libra-node",
     "libra-swarm",
     "network",
