+++ language/libra-vm/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "libra-vm"
+version = "0.1.0"
+authors = ["Libra Association <opensource@libra.org>"]
+description = "Libra VM runtime"
+repository = "https://github.com/libra/libra"
+homepage = "https://libra.org"
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+hex = "0.4.2"
+once_cell = "1.3.1"
+rayon = "1.1"
+mirai-annotations = "1.6.0"
+prometheus = { version = "0.8.0", default-features = false }
+
+bytecode-verifier = { path = "../bytecode-verifier", version = "0.1.0" }
+lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
+libra-config = { path = "../../config", version = "0.1.0" }
+libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
+libra-logger = { path = "../../common/logger", version = "0.1.0" }
+libra-metrics = { path = "../../common/metrics", version = "0.1.0" }
+libra-state-view = { path = "../../storage/state-view", version = "0.1.0" }
+libra-types = { path = "../../types", version = "0.1.0" }
+move-vm-cache = { path = "../move-vm/cache", version = "0.1.0" }
+move-vm-runtime = { path = "../move-vm/runtime", version = "0.1.0" }
+move-vm-state = { path = "../move-vm/state", version = "0.1.0" }
+move-vm-types = { path = "../move-vm/types", version = "0.1.0" }
+vm = { path = "../vm", version = "0.1.0" }
+
+[dev-dependencies]
+proptest = "0.9"
+
+[features]
+default = []
+mirai-contracts = []
+fuzzing = ["vm/fuzzing"]
