+++ language/tools/test-generation/src/lib.rs
@@ -24,6 +24,7 @@ use language_e2e_tests::executor::FakeExecutor;
 use libra_config::config::VMConfig;
 use libra_state_view::StateView;
 use libra_types::{account_address::AccountAddress, byte_array::ByteArray, vm_error::StatusCode};
+use libra_vm::LibraVM;
 use move_vm_types::values::Value;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use slog_scope::{debug, error, info};
@@ -35,7 +36,6 @@ use vm::{
     file_format::{CompiledModule, CompiledModuleMut, FunctionDefinitionIndex, SignatureToken},
     transaction_metadata::TransactionMetadata,
 };
-use vm_runtime::LibraVM;
 
 /// This function calls the Bytecode verifier to test it
 fn run_verifier(module: CompiledModule) -> Result<VerifiedModule, String> {
