+++ language/tools/vm-genesis/src/genesis_gas_schedule.rs
@@ -3,6 +3,7 @@
 
 //! This file contains the starting gas schedule published at genesis.
 
+use libra_vm::system_module_names::GAS_SCHEDULE_MODULE;
 use move_vm_runtime::MoveVM;
 use move_vm_state::{data_cache::RemoteCache, execution_context::TransactionExecutionContext};
 use move_vm_types::{loaded_data::types::Type, values::Value};
@@ -14,7 +15,6 @@ use vm::{
     },
     gas_schedule::{CostTable, GasCost, GAS_SCHEDULE_NAME, MAXIMUM_NUMBER_OF_GAS_UNITS},
 };
-use vm_runtime::system_module_names::GAS_SCHEDULE_MODULE;
 
 static INITIAL_GAS_SCHEDULE: Lazy<Vec<u8>> = Lazy::new(|| {
     use Bytecode::*;
