+++ vm-validator/Cargo.toml
@@ -19,7 +19,7 @@ scratchpad = { path = "../storage/scratchpad", version = "0.1.0" }
 libra-state-view = { path = "../storage/state-view", version = "0.1.0" }
 storage-client = { path = "../storage/storage-client", version = "0.1.0" }
 libra-types = { path = "../types", version = "0.1.0" }
-vm-runtime = { path = "../language/vm/vm-runtime", version = "0.1.0" }
+libra-vm = { path = "../language/libra-vm", version = "0.1.0" }
 
 [dev-dependencies]
 rand = "0.6.5"
@@ -28,7 +28,7 @@ config-builder = { path = "../config/config-builder", version = "0.1.0" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
 executor = { path = "../execution/executor", version = "0.1.0" }
 storage-service = { path = "../storage/storage-service", version = "0.1.0" }
-vm-runtime = { path = "../language/vm/vm-runtime", version = "0.1.0" }
+libra-vm = { path = "../language/libra-vm", version = "0.1.0" }
 transaction-builder = { path = "../language/transaction-builder", version = "0.1.0" }
 
 [features]
